Move some RPC related functionality from tl into rpc
