+++ runtime-light/stdlib/rpc/rpc-tl-query.h
@@ -9,7 +9,7 @@
 
 #include "runtime-core/class-instance/refcountable-php-classes.h"
 #include "runtime-core/runtime-core.h"
-#include "runtime-light/tl/tl-rpc-request.h"
+#include "runtime-light/stdlib/rpc/rpc-tl-request.h"
 
 struct RpcTlQuery : refcountable_php_classes<RpcTlQuery> {
   string tl_function_name;
