+++ runtime-light/tl/tl.cmake
@@ -1,6 +1,3 @@
 prepend(RUNTIME_TL_SRC ${BASE_DIR}/runtime-light/tl/
         tl-builtins.cpp
-        tl-rpc-error.cpp
-        tl-rpc-query.cpp
-        tl-rpc-request.cpp
 )
